Reorganize internal modules
