+++ packages/react-native-web/jest/setup.js
@@ -7,12 +7,12 @@ global.__DEV__ = true;
 const mockEmptyObject = {};
 
 // Make sure snapshots contain the original style objects
-jest.mock('../dist/cjs/modules/ReactNativePropRegistry', () => ({
+jest.mock('../dist/cjs/modules/StyleSheet/ReactNativePropRegistry', () => ({
   register: id => id,
   getByID: () => mockEmptyObject
 }));
 
-jest.mock('../dist/modules/ReactNativePropRegistry', () => ({
+jest.mock('../dist/modules/StyleSheet/ReactNativePropRegistry', () => ({
   register: id => id,
   getByID: () => mockEmptyObject
 }));
