+++ packages/react-native-web/src/exports/AppRegistry/renderApplication.js
@@ -12,8 +12,7 @@ import type { ComponentType } from 'react';
 
 import AppContainer from './AppContainer';
 import invariant from 'fbjs/lib/invariant';
-import hydrate from '../../modules/hydrate';
-import render from '../render';
+import render, { hydrate } from '../render';
 import styleResolver from '../StyleSheet/styleResolver';
 import React from 'react';
