+++ packages/react-native-web/src/exports/render/index.js
@@ -4,8 +4,9 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @noflow
+ * @flow
  */
 
-import { render } from 'react-dom';
+import { hydrate, render } from 'react-dom';
+export { hydrate };
 export default render;
