--- packages/react-native-web/src/modules/applyNativeMethods/index.js
@@ -1,21 +1,0 @@
-/**
- * Copyright (c) Nicolas Gallagher.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @noflow
- */
-
-import NativeMethodsMixin from '../NativeMethodsMixin';
-
-const applyNativeMethods = Component => {
-  Object.keys(NativeMethodsMixin).forEach(method => {
-    if (!Component.prototype[method]) {
-      Component.prototype[method] = NativeMethodsMixin[method];
-    }
-  });
-  return Component;
-};
-
-export default applyNativeMethods;
