--- packages/react-native-web/src/modules/hydrate/index.js
@@ -1,11 +1,0 @@
-/**
- * Copyright (c) Nicolas Gallagher.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @noflow
- */
-
-import { hydrate } from 'react-dom';
-export default hydrate;
